requirements.txt
yfinance>=0.2.38
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
statsmodels>=0.14.0
scikit-learn>=1.4.0
scipy>=1.12.0
requests>=2.31.0
python-dotenv>=1.0.0
reportlab>=4.1.0
python-bcb>=0.3.0
flask>=3.0.0
